Create README.md
# VotingSystem_App
An application that was created with Vanilla Java.
A university project that I had during the Spring semester of 2019 while being at Erasmus+ at Vilnius University in Lithuania. My team and I were asked to create an application in Java, which would help civilians to vote easily for a president during specific voting periods.